Implement basic allocator
Still need to implement deallocator

Because the kernel's scope is very minimal, the allocator is also very
minimal. I assume that there is at least 256MiB available above the
16MiB point in memory, and allocate from there. I also only allocate
contiguous pieces of memory, with the total size of the allocated memory
a multiple of 1KiB. If a large enough contiguous section of memory
cannot be found, then the kernel aborts.
This allows for an absolute maximum of 262,144 allocated objects,
assuming each object is at most 1016 bytes (to allow for the allocation
header), which should be plenty for our purposes.

Note that launching this on a system with less than ~272MiB of RAM could
easily crash the kernel/damage the hardware by trying to access memory
which does not exists. Also, running it on a system with > ~272MiB of
RAM is relatively pointless, as the kernel will be entirely unaware of
the extra RAM.