Set up and load Global Descriptor Table
Decided to go with a flat model, where the kernel has both code and data
access to the entire memory space. There is no user mode segments, as I
do not plan on supporting a user mode, and there is no task state
segment, since as I understand it that is only needed for multitasking.