Add EventLoop classes
Event loops will form the base of the different apps; at the start of
the app's main function, an event loop is created. Then, at the start of
each iteration of the event loop, a frame is fetched from the event
loop. The frame then handles any background stuff, like sleeping the
required amount at the end of the iteration so that the frame is of the
specified duration.

An event loop's `poll` method can be called during long-running
operations to handle keyboard events, so that the 256-wide ISR ring
buffer doesn't get overrun.

There are three event loop types implemented:
 1. The queued event loop consumes keyboard events into an internal
    queue that is 4x longer than the kernel queue (by default), and
    overrides earlier events when the queue gets overrun, and keeps
    track of how many events got discarded. Events are then processed
    via the event queue accessed by the `events` method, and can be
    processed either by popping from the queue until it's empty, or by
    iterating through the queue (which comes with an iterator)
 2. The callback event loop calls a callback for every keyboard event
    that is received, draining the kernel's internal keyboard event
    buffer in the process.
 3. The "ignore" event loop ignores all keyboard events, but it also
    drains the internal ISR keyboard event buffer so that when the
    program exit, it is not polluted with old events.