Snake: Instant response to input
Rather than just changing the direction the snake faces upon arrow keys,
also actually move the snake forward in that direction. If it's not a
valid direction, ignore the keypress.

It feels a bit weird when pressing the way the snake is already facing;
perhaps this should be ignored as well?

Also fix a bug where you could change the direction the snake faces
while in the help menu.